Refactor project into modules and scripts
