+++ config/term_mapping.json
